% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{fscache-package}
\alias{fscache}
\alias{fscache-package}
\title{fscache: File System Cache}
\description{
Manages a file system cache. Regular files can be moved or copied to the cache folder. Sub-folders can be created in order to organize the files. Files can be located inside the cache using a glob function. Text contents can be easily stored in and retrieved from the cache using dedicated functions. It can be used for an application or a package, as a global cache, or as a per-user cache, in which case the standard OS user cache folder will be used (e.g.: on Linux $HOME/.cache/R/my_app_or_pkg_cache_folder).
}
\details{
\emph{fscache} package.

\emph{fscache} helps you manage file caching for your application.
It creates a user cache folder at standard location, or any place you want.
You can then save text contents into new files, load contents from existing
files, move or copy files into the cache folder.
You have also the possibility to create sub-folders in order to organize your
files.
}
\seealso{
\link{Cache}.
}
\author{
\strong{Maintainer}: Pierrick Roger \email{pierrick.roger@cea.fr} (\href{https://orcid.org/0000-0001-8177-4873}{ORCID})

}
